% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metadata.R
\name{apply_dataset_metadata}
\alias{apply_dataset_metadata}
\title{Apply JSON metadata to dataframe as attributes}
\usage{
apply_dataset_metadata(.data, metadata)
}
\arguments{
\item{.data}{A Dataframe}

\item{metadata}{A list containing Dataset JSON dataset object metadata}
}
\value{
dataframe
}
\description{
This function takes supplied metadata and applies it to a dataframe as
corresponding attributes
}
\examples{
# TODO:
}
