% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_metadata.R
\name{data_metadata}
\alias{data_metadata}
\title{Create the data metadata container for a Dataset JSON object}
\usage{
data_metadata(study, metadata_version, metadata_ref)
}
\arguments{
\item{data_type}{Type of data being written. clinicalData for subject level
data, and referenceData for non-subject level data (i.e. TDMs, Associated
Persons)}
}
\value{
data_metadata object
}
\description{
Create the data metadata container for a Dataset JSON object
}
\examples{
# TODO:
}
