% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_metadata.R
\name{set_study_oid}
\alias{set_study_oid}
\alias{set_study_oid.data_metadata}
\alias{set_metadata_version}
\alias{set_metadata_ref}
\title{Set data metadata parameters}
\usage{
set_study_oid(x, study, ...)

\method{set_study_oid}{data_metadata}(x, study)

set_metadata_version(x, metadata_version, ...)

set_metadata_ref(x, metadata_ref)
}
\arguments{
\item{x}{data metadata or datasetjson object}

\item{study}{Study OID value}

\item{...}{}

\item{metadata_version}{Metadata version OID value}

\item{metadata_ref}{Metadata reference (i.e. path to Define.xml)}
}
\value{
A datasetjson or data_metadata object
}
\description{
This set of functions
}
\examples{
# TODO:
}
\concept{Data metadata setters}
