% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metadata.R
\name{dataset_metadata}
\alias{dataset_metadata}
\title{Generate an individual element that fills the itemGroupData field}
\usage{
dataset_metadata(item_id, name, label, items, .data)
}
\arguments{
\item{item_id}{Data Object ID for item in Dataset JSON object, corresponding to
ItemGroupDef/@OID in Define-XML.}

\item{name}{Dataset name}

\item{label}{Dataset Label}

\item{items}{Variable metadata}

\item{.data}{Dataframe to be written to Dataset JSON file}
}
\value{
dataset_metadata object
}
\description{
Generate an individual element that fills the itemGroupData field
}
\examples{
dataset_meta <- dataset_metadata(
  item_id = "IG.IRIS",
  name = "IRIS",
  label = "Iris",
  items = iris_items
)
}
