% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{datasetjson_example}
\alias{datasetjson_example}
\title{Get path to a datasetjson example file}
\usage{
datasetjson_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\value{
A file path string, or a character vector of available files if
\code{file} is \code{NULL}.
}
\description{
datasetjson comes bundled with sample files in its \code{inst/extdata} directory.
This function makes them easy to access.
}
\examples{
datasetjson_example()
datasetjson_example("dm.json")
}
