% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{file_metadata}
\alias{file_metadata}
\title{Create a file metadata object}
\usage{
file_metadata(data_type, version = "1.0.0")
}
\arguments{
\item{data_type}{Type of data being written. clinicalData for subject level
data, and referenceData for non-subject level data (i.e. TDMs, Associated
Persons)}

\item{version}{Dataset JSON schema version being used}
}
\value{
file_metadata object
}
\description{
Create a file metadata object
}
\examples{
# TODO:
}
