% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metadata.R
\name{gather_dataset_metadata}
\alias{gather_dataset_metadata}
\title{Gather Dataset JSON metadata from a dataframe which has Dataset JSON metadata
applied}
\usage{
gather_dataset_metadata(.data)
}
\arguments{
\item{.data}{A dataframe with Dataset JSON attributes applied}
}
\value{
A list of Dataset dataset object JSON metadata
}
\description{
This function will gather the attributes from a data frame which has Dataset
JSON metadata applied.
}
\examples{
TODO:
}
