% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_column_metadata}
\alias{get_column_metadata}
\title{Extract column metadata to data frame}
\usage{
get_column_metadata(x)
}
\arguments{
\item{x}{A datasetjson object}
}
\value{
A data frame containing the columns metadata
}
\description{
This function pulls out the column metadata from the \code{datasetjson} object
attributes into a more user-friendly data.frame.
}
\examples{

ds_json <- dataset_json(
  iris,
  item_oid = "IG.IRIS",
  name = "IRIS",
  dataset_label = "Iris",
  columns = iris_items
)

get_column_metadata(ds_json)
}
