% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_items}
\alias{iris_items}
\title{Example Variable Metadata for Iris}
\format{
\subsection{\code{iris_items} A data frame with 5 rows and 6 columns:}{

\describe{
\item{itemOID}{Unique identifier for Variable. Must correspond to ItemDef/@OID in Define-XML.}
\item{name}{Display format supports data visualization of numeric float and date values.}
\item{label}{Label for Variable}
\item{dataType}{Data type for Variable}
\item{length}{Length for Variable}
\item{keySequence}{Indicates that this item is a key variable in the dataset structure. It also provides an ordering for the keys.}
}
}
}
\usage{
iris_items
}
\description{
Example of the necessary variable metadata included in a Dataset JSON file
based on the Iris data frame.
}
\keyword{datasets}
