% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sas_date_formats}
\alias{sas_date_formats}
\title{A List of valid SAS(c) date formats}
\format{
\subsection{\code{sas_date_formats}}{

\describe{
A character vector with 45 elements
}
}
}
\usage{
sas_date_formats
}
\description{
Valid SAS(c) date formats pulled from
https://documentation.sas.com/doc/en/vdmmlcdc/8.1/ds2pg/p0bz5detpfj01qn1kz2in7xymkdl.htm
}
\keyword{datasets}
