% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schema_1_0_0}
\alias{schema_1_0_0}
\title{Dataset JSON Schema Version 1.0.0}
\format{
\subsection{\code{schema_1_0_0}}{

\describe{
A character vector with 1 element
}
}
}
\usage{
schema_1_0_0
}
\description{
This object is a character vector holding the schema for Dataset JSON Version 1.0.0
}
\keyword{datasets}
