% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schema_ndjson_1_1_0}
\alias{schema_ndjson_1_1_0}
\title{Dataset NDJSON Schema Version 1.1.0}
\format{
\subsection{\code{schema_ndjson_1_1_0}}{

\describe{
A character vector with 1 element
}
}
}
\usage{
schema_ndjson_1_1_0
}
\description{
This object is a character vector holding the schema used to validate the
metadata line of a Dataset NDJSON file. It is the CDISC-published schema
with three definitions hoisted into \verb{$defs} so that the references to them
resolve; see \code{data-raw/data.R}.
}
\keyword{datasets}
