% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{set_file_oid}
\alias{set_file_oid}
\title{Set the fileOID parameter}
\usage{
set_file_oid(dsjson, file_oid)
}
\arguments{
\item{dsjson}{datasetjson object}

\item{file_oid}{fileOID parameter, defined as "A unique identifier for this
file."}
}
\value{
datasetjson object
}
\description{
Set the fileOID parameter for a Dataset JSON object. This is automated during
\code{write_dataset_json()} using the file path unless the parameter has been set
manually.
}
\examples{
# TODO:
}
