% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metadata.R
\name{set_item_data}
\alias{set_item_data}
\title{Apply dataframe to itemData attribute}
\usage{
set_item_data(x, .data, ...)
}
\arguments{
\item{x}{Object to set itemData}

\item{.data}{Dataframe to be written to Dataset JSON file}

\item{...}{Additional params}
}
\value{
Input object with itemData applied
}
\description{
This function will set the itemData attribute within a datasetjson or
dataset_metadata object. It additionally sets the records parameter with the
proper number of rows in .data.
}
\examples{
# TODO:
}
