% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{set_originator}
\alias{set_originator}
\title{Set originator}
\usage{
set_originator(dsjson, originator)
}
\arguments{
\item{dsjson}{datasetjson object}

\item{originator}{originator parameter, defined as "The organization that
generated the Dataset-JSON file."}
}
\value{
datasetjson object
}
\description{
Set the originator metadata for a Dataset JSON object
}
\examples{
# TODO:
}
