% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{set_reference_data}
\alias{set_reference_data}
\title{Declare as reference data}
\usage{
set_reference_data(x, is_refdata)
}
\arguments{
\item{x}{datasetjson object}
}
\value{
datasetjson object
}
\description{
Sets DatasetJSON file to have the isReferenceData attribute set to TRUE
}
\examples{
ds_json <- dataset_json(iris)
ds_json <- set_as_reference_data(ds_json)
}
