% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{set_source_system}
\alias{set_source_system}
\title{Set source system information}
\usage{
set_source_system(dsjson, sys, sys_version)
}
\arguments{
\item{dsjson}{datasetjson object}

\item{sys}{sourceSystem parameter, defined as "The computer system or
database management system that is the source of the information in this
file."}

\item{sys_version}{sourceSystemVersion, defined as "The version of the
sourceSystem"}
}
\value{
datasetjson object
}
\description{
Set information about the source system used to generate the Dataset JSON
object.
}
\examples{
# TODO:
}
