% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_variable_attributes}
\alias{set_variable_attributes}
\title{Assign Dataset JSON attributes to data frame columns}
\usage{
set_variable_attributes(x)
}
\arguments{
\item{x}{A datasetjson object}
}
\value{
A datasetjson object with attributes assigned to individual
variables
}
\description{
Using the \code{columns} element of the Dataset JSON file, assign the available
metadata to individual columns
}
\examples{

ds_json <- dataset_json(
  iris,
  item_oid = "IG.IRIS",
  name = "IRIS",
  dataset_label = "Iris",
  columns = iris_items
)

ds_json <- set_variable_attributes(ds_json)
}
