% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metadata.R
\name{validate_dataset_items}
\alias{validate_dataset_items}
\title{Verify that the item metadata supplied is the appropriate format}
\usage{
validate_dataset_items(items)
}
\arguments{
\item{items}{}
}
\value{
Error Check
}
\description{
Verify that the item metadata supplied is the appropriate format
}
