% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetjson.R
\name{dataset_json}
\alias{dataset_json}
\title{Create a Dataset JSON Object}
\usage{
dataset_json(
  .data,
  file_oid = NULL,
  last_modified = NULL,
  originator = NULL,
  sys = NULL,
  sys_version = NULL,
  study = NULL,
  metadata_version = NULL,
  metadata_ref = NULL,
  item_oid = NULL,
  name = NULL,
  dataset_label = NULL,
  ref_data = FALSE,
  version = "1.1.0"
)
}
\arguments{
\item{.data}{Input data to contain within the Dataset JSON file. Written to
the itemData parameter.}

\item{file_oid}{fileOID parameter, defined as "A unique identifier for this
file."}

\item{originator}{originator parameter, defined as "The organization that
generated the Dataset-JSON file."}

\item{sys}{sourceSystem parameter, defined as "The computer system or
database management system that is the source of the information in this
file."}

\item{sys_version}{sourceSystemVersion, defined as "The version of the
sourceSystem"}

\item{study}{Study OID value}

\item{metadata_version}{Metadata version OID value}

\item{metadata_ref}{Metadata reference (i.e. path to Define.xml)}

\item{item_oid}{ID used to label dataset with the itemGroupData parameter.
Defined as "Object of Datasets. Key value is a unique identifier for
Dataset, corresponding to ItemGroupDef/@OID in Define-XML."}

\item{ref_data}{Boolean value that is set to "true" when the dataset contains
reference data (not subject data). The default value is "false".}

\item{version}{The DatasetJSON version to use. Currently only 1.1.0 is supported.}
}
\value{
dataset_json object pertaining to the specific Dataset JSON version
specific
}
\description{
Create the base object used to write a Dataset JSON file.
}
\details{
Note that DatasetJSON is on version 1.1.0. Based off findings from the pilot,
version 1.1.0 reflects feedback from the user community. Support for 1.0.0 has
been deprecated.
}
\examples{
# Create a basic object
ds_json <- dataset_json(
  iris, 
  file_oid = "/some/path",
  last_modified = "2023-02-15T10:23:15",
  originator = "Some Org", 
  sys = "source system", 
  sys_version = "1.0", 
  study = "SOMESTUDY",
  metadata_version = "MDV.MSGv2.0.SDTMIG.3.3.SDTM.1.7",
  metadata_ref = "some/define.xml",
  item_oid = "IG.IRIS",
  name = "IRIS",
  dataset_label = "Iris"
)

# Attach attributes directly
ds_json_updated <- set_data_type(ds_json, "referenceData")
ds_json_updated <- set_file_oid(ds_json_updated, "/some/path")
ds_json_updates <- set_last_modified(ds_json_updates, "2023-02-15T10:23:15")
ds_json_updated <- set_metadata_ref(ds_json_updated, "some/define.xml")
ds_json_updated <- set_metadata_version(ds_json_updated, "MDV.MSGv2.0.SDTMIG.3.3.SDTM.1.7")
ds_json_updated <- set_originator(ds_json_updated, "Some Org")
ds_json_updated <- set_source_system(ds_json_updated, "source system", "1.0")
ds_json_updated <- set_study_oid(ds_json_updated, "SOMESTUDY")
ds_json_updated <- set_item_oid(ds_json_updated, "IG.IRIS")
ds_json_updated <- set_dataset_name(ds_json_updated, "IRIS")
ds_json_updated <- set_dataset_label(ds_json_updated, "Iris")
}
