% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetjson.R
\name{dataset_json}
\alias{dataset_json}
\title{Create a Dataset JSON Object}
\usage{
dataset_json(.data, item_id, version = "1.0.0")
}
\arguments{
\item{.data}{Input data to contain within the Dataset JSON file. Written to
the itemData parameter.}

\item{item_id}{ID used to label dataset with the itemGroupData parameter.
Defined as "Object of Datasets. Key value is a unique identifier for
Dataset, corresponding to ItemGroupDef/@OID in Define-XML."}

\item{version}{Version of Dataset JSON schema to follow.}

\item{data_type}{Type of data being written. clinicalData for subject level
data, and referenceData for non-subject level data (i.e. TDMs, Associated
Persons)}
}
\value{
dataset_json object pertaining to the specific Dataset JSON version
specific
}
\description{
Create the base object used to write a Dataset JSON file.
}
\examples{
TODO: Re-look at this to allow providing the sections all separately.
# The file logicially splits into the file metadata, data metadata, and dataset metadata
}
