% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{datasetjson-package}
\alias{datasetjson}
\alias{datasetjson-package}
\title{datasetjson: Read and write CDISC Dataset JSON files}
\description{

}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraltemplate}
  \item \url{https://github.com/pharmaverse/admiraltemplate}
}

}
\concept{internal}
\keyword{internal}
