% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{datasetjson-package}
\alias{datasetjson}
\alias{datasetjson-package}
\title{datasetjson: Read and Write CDISC Dataset JSON Files}
\description{
Read, construct and write CDISC (Clinical Data Interchange Standards Consortium) Dataset JSON (JavaScript Object Notation) files, while validating per the Dataset JSON schema file, as described in CDISC (2023) \url{https://www.cdisc.org/dataset-json}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atorus-research/datasetjson}
  \item Report bugs at \url{https://github.com/atorus-research/datasetjson/issues}
}

}
\author{
\strong{Maintainer}: Mike Stackhouse \email{mike.stackhouse@atorusresearch.com} (\href{https://orcid.org/0000-0001-6030-723X}{ORCID})

Authors:
\itemize{
  \item Nicholas Masel \email{nmasel@its.jnj.com}
}

}
\concept{internal}
\keyword{internal}
