% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_json.R
\name{read_dataset_json}
\alias{read_dataset_json}
\title{Read dataset-json file with data of one SDTM/ADaM domain into data frame}
\usage{
read_dataset_json(path, version)
}
\arguments{
\item{path}{path to dataset-json file}
}
\value{
data frame with data of dataset-json file
}
\description{
Read dataset-json file with data of one SDTM/ADaM domain into data frame.
}
\examples{
\dontrun{
dm <- read_dataset_json("path/to/dm.json")
}
}
