% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_json.R
\name{read_dataset_json}
\alias{read_dataset_json}
\title{Read a Dataset JSON to datasetjson object}
\usage{
read_dataset_json(file, version)
}
\arguments{
\item{file}{File path on disk}

\item{version}{Dataset JSON schema version to validate against}
}
\value{
datasetjson object
}
\description{
This function validated a dataset JSON file on disk against the Dataset JSON schema, and if valid
returns a datasetjson object
}
\examples{

# TODO:
}
