% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_ndjson.R
\name{read_dataset_ndjson}
\alias{read_dataset_ndjson}
\title{Read a Dataset NDJSON file to datasetjson object}
\usage{
read_dataset_ndjson(file, decimals_as_floats = FALSE)
}
\arguments{
\item{file}{File path or URL of a Dataset NDJSON file, or a character string
containing NDJSON content}

\item{decimals_as_floats}{Convert variables of "decimal" type to float}
}
\value{
A dataframe with additional attributes attached containing the
DatasetJSON metadata.
}
\description{
This function reads a newline-delimited JSON (NDJSON) file following the
Dataset JSON v1.1.0 specification and returns a datasetjson object. Line 1
of the file contains the dataset metadata and each subsequent line contains
a single data row as a JSON array.
}
\details{
The resulting dataframe contains the additional metadata available on the
Dataset JSON file within the attributes to make this accessible to the user.
Note that these attributes are only populated if available.
\itemize{
\item \strong{sourceSystem}: The information system from which the content of this
dataset was source, including system name and version.
\item \strong{datasetJSONVersion}: The version of the Dataset-JSON standard used to
create the dataset.
\item \strong{fileOID}: A unique identifier for this dataset.
\item \strong{dbLastModifiedDateTime}: The date/time the source database was last
modified before creating the Dataset-JSON file.
\item \strong{originator}: The organization that generated the Dataset-JSON dataset.
\item \strong{studyOID}: Unique identifier for the study that may also function as a
foreign key to a Study/@OID in an associated Define-XML document, or to any
studyOID references that are used as keys in other documents;
\item \strong{metaDataVersionOID}: Unique identifier for the metadata version that may
also function as a foreign key to a MetaDataVersion/@OID in an associated
Define-XML file
\item \strong{metaDataRef}: URI for the metadata file describing the dataset (e.g.,
a Define-XML file).
\item \strong{itemGroupOID}: Unique identifier for the dataset that may also function
as a foreign key to an ItemGroupDef/@OID in an associated Define-XML file.
\item \strong{name}: The human-readable name for the dataset.
\item \strong{label}: A short description of the dataset.
\item \strong{columns}: An array of metadata objects that describe the dataset
variables. See \code{dataset_json()} for further information on the contents of
these fields.
}
}
\examples{
# Read from disk
\dontrun{
  dat <- read_dataset_ndjson("path/to/file.ndjson")
}

# Read from an already imported character vector
ds_json <- dataset_json(iris, "IG.IRIS", "IRIS", "Iris", columns=iris_items)
ndjson <- write_dataset_ndjson(ds_json)
dat <- read_dataset_ndjson(ndjson)
}
