% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schema_ndjson_1_1_0}
\alias{schema_ndjson_1_1_0}
\title{Dataset NDJSON Schema Version 1.1.0}
\format{
\subsection{\code{schema_ndjson_1_1_0}}{

\describe{
A character vector with 1 element
}
}
}
\usage{
schema_ndjson_1_1_0
}
\description{
This object is a character vector holding the schema for Dataset NDJSON
Version 1.1.0, used to validate the metadata line of NDJSON files.
}
\keyword{datasets}
