% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataset_ndjson.R
\name{write_dataset_ndjson}
\alias{write_dataset_ndjson}
\title{Write out a Dataset JSON file in NDJSON format}
\usage{
write_dataset_ndjson(
  x,
  file,
  pretty = FALSE,
  float_as_decimals = FALSE,
  digits = 16
)
}
\arguments{
\item{x}{datasetjson object}

\item{file}{File path to save NDJSON file. If omitted, the NDJSON content is
returned as a character string.}

\item{pretty}{If TRUE, write the metadata line with readable formatting.
Data row lines are always compact. \emph{Note: The Dataset JSON standard prefers
compressed formatting without line feeds. It is not recommended you use
pretty printing for submission purposes.}}

\item{float_as_decimals}{If TRUE, Convert float variables to "decimal" data
type in the JSON output. This will manually convert the numeric values
using the \code{format()} function using the number of digits specified in
\code{digits}, bypassing the \code{yyjsonr} handling of float values and writing the
numbers out as JSON character strings. See the \href{https://wiki.cdisc.org/display/PUB/Precision+and+Rounding}{Dataset JSON user guide} for more
information. Defaults to FALSE}

\item{digits}{When using \code{float_as_decimals}, the number of digits to use
when writing out floats. Going higher than 16 may start writing otherwise
sufficiently precise decimals (i.e. .2) to long strings.}
}
\value{
NULL when file written to disk, otherwise character string
}
\description{
Write a datasetjson object to a newline-delimited JSON (NDJSON) file. Line 1
contains the dataset metadata as a JSON object, and each subsequent line
contains a single data row as a JSON array.
}
\examples{
# Write to character object
ds_json <- dataset_json(
  iris,
  item_oid = "IG.IRIS",
  name = "IRIS",
  dataset_label = "Iris",
  columns = iris_items
)
ndjson <- write_dataset_ndjson(ds_json)

# Write to disk
\dontrun{
  write_dataset_ndjson(ds_json, "path/to/file.ndjson")
}
}
